requirements.txt
absl-py==1.4.0
accelerate==0.21.0
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
anthropic==0.3.9
anyio==3.7.0
appdirs==1.4.4
aspy.yaml==1.3.0
astor==0.8.1
astroid==2.5.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==20.3.0
backcall==0.2.0
bitsandbytes==0.39.1
black==23.3.0
blessed==1.20.0
blinker==1.6.2
cached-property==1.5.2
cachetools==5.3.1
certifi==2020.12.5
cfgv==3.2.0
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.1.0
cpm-kernels==1.0.11
cycler==0.11.0
datasets==2.10.1
decorator==4.4.2
deepspeed==0.10.0
dill==0.3.4
distlib==0.3.1
distro==1.8.0
docker-pycreds==0.4.0
einops==0.6.1
evaluate==0.4.0
exceptiongroup==1.1.1
executing==1.2.0
fastapi==0.98.0
ffmpy==0.3.0
filelock==3.0.12
fire==0.5.0
flash-attn==1.0.4
Flask==2.3.2
Flask-Cors==4.0.0
fonttools==4.40.0
frozenlist==1.3.3
fsspec==2023.6.0
ftfy==6.1.1
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.22.0
google-auth-oauthlib==1.0.0
gpustat==1.1
gradio==3.23.0
grpcio==1.56.2
h11==0.14.0
hjson==3.1.0
httpcore==0.17.2
httpx==0.24.1
huggingface-hub==0.15.1
icetk==0.0.7
identify==2.2.0
idna==2.10
importlib-metadata==6.8.0
importlib-resources==5.12.0
iniconfig==1.1.1
ipykernel==4.6.0
ipython==8.13.0
ipython-genutils==0.2.0
isort==5.8.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.17.3
jupyter-client==6.1.12
jupyter-core==4.7.1
kiwisolver==1.4.4
lazy-object-proxy==1.5.2
linkify-it-py==2.0.2
loguru==0.7.0
loralib==0.1.1
Markdown==3.4.4
markdown-it-py==2.2.0
markdown2==2.4.9
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mccabe==0.6.1
mdit-py-plugins==0.3.3
mdurl==0.1.2
mpi4py==3.1.4
msgpack==1.0.5
multidict==6.0.4
multiprocess==0.70.12.2
mypy-extensions==1.0.0
nh3==0.2.13
ninja==1.11.1
nodeenv==1.5.0
numpy==1.24.2
nvidia-ml-py==12.535.77
oauthlib==3.2.2
openai==0.27.8
orjson==3.9.1
packaging==23.1
pandas==2.0.3
parso==0.8.3
pathspec==0.11.1
pathtools==0.1.2
peft==0.4.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.2
pkgutil_resolve_name==1.3.10
platformdirs==3.8.0
pluggy==0.13.1
pre-commit==1.10.4
prompt-toolkit==3.0.38
protobuf==3.20.0
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.10.0
py-cpuinfo==9.0.0
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.9
pydub==0.25.1
pyg-lib==0.2.0
Pygments==2.15.1
PyGObject==3.26.1
pylint==2.7.2
pyparsing==2.4.7
pyrsistent==0.19.3
pytest==6.2.2
python-apt==1.6.5+ubuntu0.7
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3
PyYAML==5.4.1
pyzmq==22.0.3
ray==2.6.1
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
responses==0.18.0
rich==13.4.2
rsa==4.9
safetensors==0.3.1
scikit-learn==1.2.2
scipy==1.10.1
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==1.26.0
setproctitle==1.3.2
shortuuid==1.0.11
simplegeneric==0.8.1
six==1.15.0
smmap==5.0.0
sniffio==1.3.0
ssh-import-id==5.7
stack-data==0.6.2
starlette==0.27.0
svgwrite==1.4.3
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorboardX==2.6.1
termcolor==2.3.0
threadpoolctl==3.2.0
tiktoken==0.4.0
tokenize-rt==5.1.0
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
toolz
# torch==1.13.0
torch-cluster==1.6.1
torch-geometric==2.3.1
torch-scatter==2.1.0
torch-sparse==0.6.16
torch-spline-conv==1.2.2
# torchvision==0.14.0
tornado==6.1
tqdm
traitlets==5.0.5
transformers==4.31.0
trl
typing_extensions==4.7.0
tzdata==2023.3
uc-micro-py==1.0.2
unattended-upgrades==0.1
urllib3
uvicorn==0.22.0
virtualenv==20.4.3
wandb==0.15.11
wavedrom==2.0.3.post3
wcwidth==0.2.5
websockets==11.0.3
Werkzeug==2.3.6
wrapt==1.12.1
xxhash==3.2.0
yarl==1.9.2
zipp==3.15.0
